+++ src/components/base/ui/MxPopover.vue
@@ -8,7 +8,7 @@
 				v-bind="content"
 				avoid-collisions
 				position-strategy="absolute"
-				class="bg-ui-menu-bg-inner rounded p-2 z-1000 text-ui-menu-bg-text text-sm"
+				class="bg-ui-menu-bg-inner rounded p-2 z-3000 text-ui-menu-bg-text text-sm"
 				@open-auto-focus.prevent
 			>
 				<slot name="content"></slot>