+++ src/components/preferences/sections/themes/blocks/number-field.ts
@@ -1,0 +1,57 @@
+import type { ThemeInputField } from '../types'
+
+export const ThemeFieldsNumberField: ThemeInputField[] = [
+	{
+		key: '--color-ui-number-text',
+		type: 'color',
+		label: 'Text',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-text-focused',
+		type: 'color',
+		label: 'Text Focused',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-selection',
+		type: 'color',
+		label: 'Selection',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-slider',
+		type: 'color',
+		label: 'Slider',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-bg',
+		type: 'color',
+		label: 'Background',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-bg-focused',
+		type: 'color',
+		label: 'Background Focused',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-outline',
+		type: 'color',
+		label: 'Outline',
+		transparent: true
+	},
+	{
+		key: '--color-ui-number-outline-focused',
+		type: 'color',
+		label: 'Outline Focused',
+		transparent: true
+	},
+	{
+		key: '--ui-number-roundness',
+		type: 'number',
+		label: 'Roundness'
+	}
+]