+++ src/components/preferences/sections/themes/types.ts
@@ -1,0 +1,14 @@
+export type ThemeInputField = ThemeInputFieldColor | ThemeInputFieldNumber
+
+type ThemeInputFieldColor = {
+	type: 'color'
+	key: string
+	label: string
+	transparent?: boolean
+}
+
+type ThemeInputFieldNumber = {
+	type: 'number'
+	key: string
+	label: string
+}